refactor: split node storage api routes
